pyproject.toml
[tool.poetry]
name = "Bitbucket"
version = "0.1.0"
description = "Bitbucket"
authors = ["AnnaKPolyakova"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.11"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.22.0"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
isort = "^5.10.1"
black = "^22.3.0"
django-filter = "^21.1"
pytest-cov = "^3.0.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-environ = "^0.8.1"
django-debug-toolbar = "^3.2.4"
django-cors-headers = "^3.11.0"
python-dotenv = "^0.20.0"
geopy = "^2.2.0"
factory-boy = "^3.2.1"
pytz = "^2022.1"
pytils = "^0.4.1"
celery = "^5.2.6"
apachelogs = "^0.6.0"
redis = "^4.2.2"
django-admin-rangefilter = "^0.8.4"
django-admin-list-filter-dropdown = "^1.0.3"

